requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.1
django-allauth==0.50.0
django-autoslug==1.9.8
django-cities==0.6
django-cities-light==3.9.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-environ==0.8.1
django-model-utils==4.2.0
django-sslserver==0.22
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
progressbar2==4.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pypaystack==1.24
python-utils==3.1.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2021.5
Unidecode==1.3.4
urllib3==1.26.9
